+++ src/renderer/ipcRequest/app.ts
@@ -8,3 +8,21 @@ export function openExternalUrl(url: string) {
 export function requestResetApp() {
   window.rabbyDesktop.ipcRenderer.sendMessage('__internal_rpc:app:reset-app');
 }
+
+export function getPersistedProxyConfig() {
+  return window.rabbyDesktop.ipcRenderer.invoke('get-proxyConfig');
+}
+
+export function applyProxyConfig(conf: IAppProxyConf) {
+  return window.rabbyDesktop.ipcRenderer.invoke('apply-proxyConfig', conf);
+}
+
+export function validateProxyConfig(
+  url: string,
+  proxyConfig: IAppProxyConf['proxySettings']
+) {
+  return window.rabbyDesktop.ipcRenderer.invoke('check-proxyConfig', {
+    detectURL: url,
+    proxyConfig,
+  });
+}